requirements.txt
pika==0.10.0 \
    --hash=sha256:63131aaeec48a6c8f1db1fe657e1e74cf384c3927eb7d1725e31edae4220dea4 \
    --hash=sha256:7277b4d12a99efa4058782614d84138983f9f89d690bdfcea66290d810806459
requests==2.18.3 \
    --hash=sha256:b62be4ec5999c24d10c98d248a136e7db20ca6616a2b65060cd9399417331e8a \
    --hash=sha256:fb68a7baef4965c12d9cd67c0f5a46e6e28be3d8c7b6910c758fbcc99880b518
six==1.10.0 \
    --hash=sha256:0ff78c403d9bccf5a425a6d31a12aa6b47f1c21ca4dc2573a7e2f32a97335eb1 \
    --hash=sha256:105f8d68616f8248e24bf0e9372ef04d3cc10104f1980f54d57b2ce73a5ad56a
structlog==16.1.0 \
    --hash=sha256:a8cc45c208c6b251031ec223940552eccd97dbdb322474e0c4ec9e50f76a225a \
    --hash=sha256:b44dfaadcbab84e6bb97bd9b263f61534a79611014679757cd93e2359ee7be01
sdx-common==0.7.1 \
    --hash=sha256:8327955891185d310259998c2a36504abc3ad24d3939397645a9df20b27fd4b8 \
    --hash=sha256:1dadd670763e0e4ef622db4a4ad84ed302e57143ebc6543f3e832654c8d8ee4e
sdc-rabbit==0.3.4 \
    --hash=sha256:4d0ba9ba0aedd8779de388076eaac961d5ae74a5af3f381f9f8590978988a540 \
    --hash=sha256:90fecd891fbcc48c5afe3323c5f1056f6313d069c65bdfdc8b6c9c65e4d48f80
chardet==3.0.4 \
    --hash=sha256:fc323ffcaeaed0e0a02bf4d117757b98aed530d9ed4531e3e15460124c106691 \
    --hash=sha256:84ab92ed1c4d4f16916e05906b6b75a6c0fb5db821cc65e70cbd64a3e2a5eaae
urllib3==1.22 \
    --hash=sha256:06330f386d6e4b195fbfc736b297f58c5a892e4440e54d294d7004e3a9bbea1b \
    --hash=sha256:cc44da8e1145637334317feebd728bd869a35285b93cbb4cca2577da7e62db4f
certifi==2017.7.27.1 \
    --hash=sha256:54a07c09c586b0e4c619f02a5e94e36619da8e2b053e20f594348c0611803704 \
    --hash=sha256:40523d2efb60523e113b44602298f0960e900388cf3bb6043f645cf57ea9e3f5
idna==2.6 \
    --hash=sha256:8c7309c718f94b3a625cb648ace320157ad16ff131ae0af362c9f21b80ef6ec4 \
    --hash=sha256:2c6a5de3089009e3da7c5dde64a141dbc8551d5b7f6cf4ed7c2568d0cc520a8f